Separated watch 3d node into 2 projects.
