+++ src/Libraries/Watch3DNodeModelsWpf/Properties/AssemblyInfo.cs
