+++ src/Libraries/Watch3DNodeModelsWpf/packages.config
